9 React moved updateCat to singleCat
